Time: 79 ms (22.27%) | Memory: 193.6 MB (17.32%) - LeetSync

+++ 943-sum-of-subarray-minimums/sum-of-subarray-minimums.cpp
@@ -39,6 +39,6 @@ public:
 
             total = (total + (rgt * left * 1LL * arr[i]) % MOD) % MOD;
         }
-        return total;
+        return (int)total;
     }
 };
